reshuffle publish/auth tag interfaces

+++ src/main/java/com/tlabs/eve/esi/ESIService.java
@@ -40,7 +40,7 @@ public interface ESIService {
 
     Map<Long, ESILocation.SolarSystem> getSolarSystemStatistics();
 
-    List<ESIName> getNames(final List<Long> ids);
+    List<ESIName> getNames(final List<Integer> ids);
 
     ESIKillMail getKillMail(final ESIKillMail killMail);
